Added checkout_create method
Either mode is 'register' and an account wil be created immediately or mode is 'guest' and a guest order will be made or if there is a match for email, first_name, last_name and phone an existing customer will be attached to the given order. In both cases these fields are required.